refactor: #28 - Feature 모듈 재구성 및 라우팅 구조 개선
- FeatureHome에서 Group, Experience를 독립 모듈로 분리
  (FeatureHome은 Home 라우팅 허브만 담당)
- FeatureReport 제거 → ReportView를 FeatureMain으로 흡수
- OnboardingView의 NavigationStack/destination dispatcher를
  App 타겟의 RootView로 이전 (FeatureAuth가 다른 Feature를 의존하지 않도록)
- App 타겟에 ContentView 분리 (로그인/온보딩 분기)
- 모듈 외부에 노출되는 View들에 public 처리
- FeatureMain, FeatureAuth에 DotLottie 직접 의존성 추가
