--- Sources/FeatureMain/ContentView.swift
@@ -1,37 +1,0 @@
-//
-//  ContentView.swift
-//  WatchOut
-//
-//  Created by 어재선 on 9/2/25.
-//
-
-import SwiftUI
-import Data
-import Shared
-import FeatureAuth
-
-struct ContentView: View {
-    @StateObject private var appState = AppState()
-    @StateObject private var userManger = UserManager.shared
-    
-    var body: some View {
-        
-        VStack {
-            if !appState.isLoggedIn {
-        
-                LoginView()
-                    .navigationBarBackButtonHidden()
-                    .environmentObject(appState)
-            } else {
-                OnboardingView()
-                    .navigationBarBackButtonHidden()
-                    .environmentObject(appState)
-            }
-        }
-        .environmentObject(userManger)
-    }
-}
-
-#Preview {
-    ContentView()
-}
